remove: dependency of cookies in middleware to prevent sensible data
