+++ utils/constants.ts
@@ -1,0 +1,2 @@
+export const API_URL = process.env.NEXT_PUBLIC_API_URL;
+export const API_KEY = process.env.NEXT_PUBLIC_API_KEY;
